+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/dto/ProductRatingResponse.java
@@ -7,10 +7,10 @@ import java.util.Date;
 
 @Data
 public class ProductRatingResponse {
-    private short id;
-    private short authorId;
+    private Long id;
+    private Long authorId;
     private String author;
-    private short rating;
+    private int rating;
     private Product product;
     private String message;
     private boolean approved;
