+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/dto/auth/AuthenticationResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder
 public class AuthenticationResponse {
-    private short id;
+    private Long id;
     private String email;
     private String token;
     private String refreshToken;
