+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/dto/category/CategoryResponse.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Data
 public class CategoryResponse {
-    private short id;
+    private Long id;
     private String name;
     private String description;
     private String imageURL;
