+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/exception/EntityNotFoundException.java
@@ -4,7 +4,7 @@ public class EntityNotFoundException extends RuntimeException{
     public EntityNotFoundException(String msg) {
         super(msg);
     }
-    public EntityNotFoundException(String className, Short id) {super(className + " with id " + id + " not found");}
+    public EntityNotFoundException(String className, Long id) {super(className + " with id " + id + " not found");}
     public EntityNotFoundException(String msg, Throwable cause) {
         super(msg, cause);
     }
