+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/mapper/CategoryMapper.java
@@ -27,7 +27,7 @@ public class CategoryMapper {
                 .collect(Collectors.toList());
     }
 
-    public CategoryResponse getCategoryById(Short id) {
+    public CategoryResponse getCategoryById(Long id) {
         return mapToResponseDTO(service.getCategoryById(id));
     }
