+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/repository/ProductRatingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRatingRepository extends JpaRepository<ProductRating, Short> {
+public interface ProductRatingRepository extends JpaRepository<ProductRating, Long> {
 }
