+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/CategoryService.java
@@ -6,9 +6,9 @@ import by.dz.fruits_in_choco.fruits_in_choco.entity.category.Category;
 import java.util.List;
 
 public interface CategoryService {
-    Category getCategoryById(short id);
+    Category getCategoryById(Long id);
     List<Category> getCategories(int page, int size, String direction, String sortBy);
     Category saveCategory(CategoryRequest request);
-    Category updateCategory(Category category, short id);
-    void deleteCategoryById(short id);
+    Category updateCategory(Category category, Long id);
+    void deleteCategoryById(Long id);
 }
