+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/OrderService.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Map;
 
 public interface OrderService {
-    Order makeOrder(Order order, Map<Short, Short> products);
+    Order makeOrder(Order order, Map<Long, Integer> products);
     List<Order> getAllOrders();
-    Order getOrderById(short id);
-    void deleteOrderById(short id);
-    Order updateOrder(Order order, short id);
+    Order getOrderById(Long id);
+    void deleteOrderById(Long id);
+    Order updateOrder(Order order, Long id);
 }
