+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/ProductService.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getProducts(int page, int size, String direction, String sortBy);
-    Product getProductById(short id);
-    List<Product> getProductsFilteredByCategories(List<Short> categories);
+    Product getProductById(Long id);
+    List<Product> getProductsFilteredByCategories(List<Long> categories);
     Product saveProduct(Product product);
-    Product updateProduct(Product product, short id);
-    void deleteProductById(short id);
-    Product rateProduct(ProductRatingRequest request, short id);
-    Product approveReview(ProductRatingRequest rating, short productId, short ratingId);
-    void deleteProductRatingById(short productId, short ratingId);
+    Product updateProduct(Product product, Long id);
+    void deleteProductById(Long id);
+    Product rateProduct(ProductRatingRequest request, Long id);
+    Product approveReview(ProductRatingRequest rating, Long productId, Long ratingId);
+    void deleteProductRatingById(Long productId, Long ratingId);
 }
